JTable toegevoegd met historie, bugfixes, werkt met echte traces, enkel doet victim cache blijkbaar nog niets...

+++ SImulator/src/simulator/parts/Cache.java
@@ -77,6 +77,6 @@ public abstract class Cache {
      * @param address
      * @return
      */
-    public abstract boolean access(int address);
+    public abstract boolean access(long address);
 
 }
